+++ addons/godot_dotnet_mcp/plugin/runtime/plugin_runtime_state.gd
@@ -39,6 +39,7 @@ const DEFAULT_COLLAPSED_SYSTEM_TOOLS: Array[String] = [
 	"system_plugin_reload",
 	"system_plugin_update",
 	"system_project_configure",
+	"system_project_execution",
 	"system_project_files",
 	"system_project_index_build",
 	"system_project_run",